+++ tests/Test/Functional/DataAccessTest.php
@@ -58,8 +58,8 @@ class DataAccessTest extends FunctionalTestCase
         $sql  = 'SELECT test_int, test_string FROM ' . $this->table . ' WHERE test_int = ? AND test_string = ?';
         $stmt = $this->connection->prepare($sql);
 
-        $stmt->bindParam(1, $paramInt, ParameterType::INTEGER);
-        $stmt->bindParam(2, $paramStr, ParameterType::STRING);
+        $stmt->bindValue(1, $paramInt, ParameterType::INTEGER);
+        $stmt->bindValue(2, $paramStr, ParameterType::STRING);
 
         $row = $stmt->execute()->fetchAssociative();
 
@@ -76,8 +76,8 @@ class DataAccessTest extends FunctionalTestCase
         $sql  = 'SELECT test_int, test_string FROM ' . $this->table . ' WHERE test_int = ? AND test_string = ?';
         $stmt = $this->connection->prepare($sql);
 
-        $stmt->bindParam(1, $paramInt, ParameterType::INTEGER);
-        $stmt->bindParam(2, $paramStr, ParameterType::STRING);
+        $stmt->bindValue(1, $paramInt, ParameterType::INTEGER);
+        $stmt->bindValue(2, $paramStr, ParameterType::STRING);
 
         $rows    = $stmt->execute()->fetchAllAssociative();
         $rows[0] = array_change_key_case($rows[0], CASE_LOWER);
@@ -92,8 +92,8 @@ class DataAccessTest extends FunctionalTestCase
         $sql  = 'SELECT test_int FROM ' . $this->table . ' WHERE test_int = ? AND test_string = ?';
         $stmt = $this->connection->prepare($sql);
 
-        $stmt->bindParam(1, $paramInt, ParameterType::INTEGER);
-        $stmt->bindParam(2, $paramStr, ParameterType::STRING);
+        $stmt->bindValue(1, $paramInt, ParameterType::INTEGER);
+        $stmt->bindValue(2, $paramStr, ParameterType::STRING);
 
         $column = $stmt->execute()->fetchOne();
         self::assertSame(1, $column);
@@ -106,7 +106,9 @@ class DataAccessTest extends FunctionalTestCase
 
         $sql    = 'SELECT test_int, test_string FROM ' . $this->table . ' WHERE test_int = ? AND test_string = ?';
         $stmt   = $this->connection->prepare($sql);
-        $result = $stmt->execute([$paramInt, $paramStr]);
+        $stmt->bindValue(1, $paramInt, ParameterType::INTEGER);
+        $stmt->bindValue(2, $paramStr, ParameterType::STRING);
+        $result = $stmt->execute();
 
         $row = $result->fetchAssociative();
         self::assertNotFalse($row);
@@ -387,6 +389,7 @@ class DataAccessTest extends FunctionalTestCase
         $table->setPrimaryKey(['test_date']);
 
         $sm = $this->connection->createSchemaManager();
+        $this->dropTableIfExists('fetch_table_date_math');
         $sm->createTable($table);
 
         $this->connection->insert('fetch_table_date_math', ['test_date' => '2010-01-01', 'test_days' => 10]);